backend/requirements.txt
fastapi
uvicorn
requests
pydantic
email-validator
python-dotenv
